Major refactor of Higad class from Game
