+++ src/constants.js
@@ -16,5 +16,6 @@ module.exports = Object.freeze({
         DIRECTION_DOWN: DIRECTION_UP,
         DIRECTION_LEFT: DIRECTION_RIGHT,
         DIRECTION_RIGHT: DIRECTION_LEFT,
-    }
+    },
+    DIRECTIONS: [DIRECTION_UP, DIRECTION_DOWN, DIRECTION_LEFT, DIRECTION_RIGHT],
 })
